requirements.txt
openai-agents>=0.2.11
click>=8.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
openai>=1.0.0